# perf(http): remove brand check in `Response` constructor