Code:

```js
Deno.serve({ port: 8085 }, (request) => {
  return new Response(request.url);
});
```

Before:

```
$ wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    64.34us   34.35us   3.32ms   96.73%
    Req/Sec    76.42k     2.16k   80.22k    79.45%
  9138597 requests in 1.00m, 1.37GB read
Requests/sec: 152057.54
Transfer/sec:     23.35MB
```

After:

```
$ wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    64.25us   54.13us   5.78ms   98.55%
    Req/Sec    77.07k     2.04k   80.99k    72.71%
  9218839 requests in 1.00m, 1.38GB read
Requests/sec: 153392.88
Transfer/sec:     23.55MB
```