@dsherret 2024-10-14
I'm going to close this one because using webidl.createBranded elsewhere led to a slowdown.